user delete todo comments

+++ models/userModel.js
@@ -48,7 +48,9 @@ exports.updateUser = async (uid, data) => {
 
 exports.deleteUser = async (uid) => {
     try {
-        const result = await db.sql`DELETE FROM users WHERE uid = ${uid} RETURNING *`;
+        //delete all items Use Join
+        //delete all bags with uid
+        const result = await db.sql`DELETE FROM users WHERE uid = ${uid} RETURNING *`; //delete user with uid
         return result[0];
     } catch(error) {
         console.log(error);
